refactor(samples): split samples into their own files
